<feat> Add styling to current section for mobile
When viewing the page on a device with a screen width of less than
750px, the information is organized using flexbox in a column orientation.
I found that the current section was lost or under-emphasized.  To
counter this I added the node styling as bullet points and changed the
styling of the text to use the theme color scheme. The color also added
a cohesiveness when viewing on a desktop as the legend and title of the
skills graph also uses the same color.