<fix> Move carousel varible update to ngOnChanges
Found an error when selecting a project with no images, then selecting
a project with images.  The index that manages which images are selected
from the img array was set to the previous project rather than the
selected one.  By moving the variable update to ngOnChanges, the app now
updates those values each time the user selects a project.