<style> Change Style Values on grid and font color
Found an issue with the grid in the contact view.  The grid was using
a view height value to generate space for the form and the image.  Since
the view height was different between a cellphone and my desktop, there
was an uncomfortable amount of space between the image and the form.
I fixed this by changing the value to a rem measurement.

Also updated the color of the required notation in the form.  The value
was using the dark variation of the accent theme.  This was ledgable on
my workstaion which has white washed monitors but I found on other devices
that the font was not ledgable.  Changed to a light variant of the
accent theme.